vim: vgit: gj/gJ jumps to next alphabetical hunk
Instead of always jumping to the first or last hunk.
This is useful if you are purposefully skipping a file or
two in the qf list, for example, while an AI is working on fixing them.
You want to proceed down the list. After staging all hunks in a file,
the automatic gj from quitting vgit takes you back to the first file
again (which may not yet be ready for review). Instead, we will look at
where our current file is ordered alphabetically amongst all unstaged
hunks, then jump to the next or prev hunk according to this alphabetical
ordering.